Bottom Shett Dialog 클래스 분리
