+++ app/src/main/java/com/myongsik/myongsikandroid/ui/view/food/MapDetailBottomSheetDialog.kt
@@ -1,0 +1,150 @@
+package com.myongsik.myongsikandroid.ui.view.food
+
+import android.content.Intent
+import android.net.Uri
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.Toast
+import androidx.fragment.app.viewModels
+import com.google.android.material.bottomsheet.BottomSheetBehavior
+import com.google.android.material.bottomsheet.BottomSheetDialogFragment
+import com.myongsik.myongsikandroid.R
+import com.myongsik.myongsikandroid.data.model.food.RequestScrap
+import com.myongsik.myongsikandroid.data.model.kakao.Restaurant
+import com.myongsik.myongsikandroid.databinding.DialogBottomRestaurantSheetBinding
+import com.myongsik.myongsikandroid.ui.viewmodel.food.HomeViewModel
+import com.myongsik.myongsikandroid.ui.viewmodel.search.LoveViewModel
+import com.myongsik.myongsikandroid.util.MyongsikApplication
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MapDetailBottomSheetDialog(private val restaurantId: Int) : BottomSheetDialogFragment() {
+
+    private var _binding : DialogBottomRestaurantSheetBinding? = null
+    private val binding : DialogBottomRestaurantSheetBinding
+        get() = _binding!!
+
+    private val homeViewModel by viewModels<HomeViewModel>()
+    private val loveViewModel by viewModels<LoveViewModel>()
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = DialogBottomRestaurantSheetBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        val bottomSheetBehavior = BottomSheetBehavior.from(view.parent as View)
+        bottomSheetBehavior.addBottomSheetCallback(object : BottomSheetBehavior.BottomSheetCallback() {
+            override fun onStateChanged(bottomSheet: View, newState: Int) {
+                if (newState == BottomSheetBehavior.STATE_HIDDEN) {
+                    dismiss()
+                }
+            }
+
+            override fun onSlide(bottomSheet: View, slideOffset: Float) {
+            }
+        })
+        homeViewModel.getOneRestaurant(restaurantId)
+        initView()
+    }
+
+    private fun initView() {
+        val restaurantScrapCount = binding.dialogScrapCountNumberTv
+        val restaurantCategory = binding.dialogCategoryTv
+        val restaurantAddress = binding.dialogAddressTv
+        val restaurantCallIv = binding.dialogBottomCallIv
+        val restaurantScrapIv = binding.dialogBottomLoveIv
+        val restaurantScrapOnIv = binding.dialogBottomScrapIv
+        var restaurantPhoneNum = ""
+
+        var restaurantScrap : RequestScrap? = null
+        var restaurant : Restaurant? = null
+        var scrapCount = 0
+
+        homeViewModel.getDetailRestaurant.observe(viewLifecycleOwner){
+            binding.dialogNameTv.text = it.data.name
+            scrapCount = it.data.scrapCount!!.toInt()
+            restaurantScrapCount.text = it.data.scrapCount.toString()
+            restaurantCategory.text = it.data.category
+            restaurantAddress.text = it.data.address
+            restaurantPhoneNum = it.data.contact
+            //스크랩
+            restaurantScrap = RequestScrap(
+                address = it.data.address,
+                campus = MyongsikApplication.prefs.getUserCampus(),
+                category = it.data.category,
+                code = it.data.code,
+                contact = it.data.contact,
+                distance = it.data.distance,
+                name = it.data.name,
+                phoneId = MyongsikApplication.prefs.getUserID(),
+                urlAddress = it.data.urlAddress,
+                latitude = it.data.latitude.toString(),
+                longitude = it.data.longitude.toString()
+            )
+            restaurant = Restaurant(
+                address_name = it.data.address,
+                category_group_code = " ",
+                category_group_name = it.data.category,
+                category_name = it.data.category,
+                distance = it.data.distance,
+                id = it.data.code,
+                phone = it.data.contact,
+                place_name = it.data.name,
+                place_url = it.data.urlAddress,
+                road_address_name = it.data.address,
+                x = it.data.latitude.toString(),
+                y = it.data.longitude.toString()
+            )
+            loveViewModel.loveIs(restaurant!!)
+        }
+
+        loveViewModel.loveIs.observe(viewLifecycleOwner){
+            if(it == null){
+                restaurantScrapIv.visibility = View.VISIBLE
+                restaurantScrapOnIv.visibility = View.INVISIBLE
+            }
+            else{
+                restaurantScrapIv.visibility = View.INVISIBLE
+                restaurantScrapOnIv.visibility = View.VISIBLE
+            }
+        }
+
+        restaurantCallIv.setOnClickListener {
+            restaurantClick(restaurantPhoneNum)
+        }
+
+        restaurantScrapIv.setOnClickListener {
+            loveViewModel.scarpRestaurant(restaurantScrap!!)
+            loveViewModel.saveFoods(restaurant!!)
+            Toast.makeText(requireContext(), "찜 완료!", Toast.LENGTH_SHORT).show()
+            restaurantScrapCount.text = (scrapCount+1).toString()
+            restaurantScrapOnIv.visibility = View.VISIBLE
+            restaurantScrapIv.visibility = View.INVISIBLE
+        }
+
+        restaurantScrapOnIv.setOnClickListener {
+            loveViewModel.deleteFoods(restaurant!!)
+            Toast.makeText(requireContext(), "찜 목록에서 삭제되었습니다.", Toast.LENGTH_SHORT).show()
+            restaurantScrapOnIv.visibility = View.INVISIBLE
+            restaurantScrapIv.visibility = View.VISIBLE
+        }
+    }
+
+    private fun restaurantClick(phoneNum : String) {
+        if (context?.getString(R.string.is_null_phone_number) != phoneNum) {
+            val intent = Intent(Intent.ACTION_VIEW, Uri.parse("tel:$phoneNum"))
+            context?.startActivity(intent)
+        } else {
+            Toast.makeText(requireContext(), getString(R.string.is_null_phone_number), Toast.LENGTH_SHORT).show()
+        }
+    }
+}
